+++ traffic_ops/testing/api/v3/cdn_domains_test.go
@@ -16,10 +16,11 @@ package v3
 */
 
 import (
-	"github.com/apache/trafficcontrol/lib/go-rfc"
 	"net/http"
 	"testing"
 	"time"
+
+	"github.com/apache/trafficcontrol/lib/go-rfc"
 )
 
 func GetTestDomains(t *testing.T) {