+++ traffic_ops/traffic_ops_golang/coordinate/coordinates.go
@@ -50,9 +50,13 @@ func (v *TOCoordinate) ParamColumns() map[string]dbhelpers.WhereColumnInfo {
 		"name": dbhelpers.WhereColumnInfo{"name", nil},
 	}
 }
+
+func (v *TOCoordinate) GetLastUpdated() (*time.Time, bool, error) {
+	return api.GetLastUpdated(v.APIInfo().Tx, *v.ID, "coordinate")
+}
+
 func (v *TOCoordinate) UpdateQuery() string { return updateQuery() }
 func (v *TOCoordinate) DeleteQuery() string { return deleteQuery() }
-
 func (coordinate TOCoordinate) GetKeyFieldsInfo() []api.KeyFieldInfo {
 	return []api.KeyFieldInfo{{"id", api.GetIntKey}}
 }
@@ -131,7 +135,9 @@ func (v *TOCoordinate) SelectMaxLastUpdatedQuery(where, orderBy, pagination, tab
 	select max(last_updated) as t from last_deleted l where l.table_name='` + tableName + `') as res`
 }
 
-func (coord *TOCoordinate) Update() (error, error, int) { return api.GenericUpdate(coord) }
+func (coord *TOCoordinate) Update(h http.Header) (error, error, int) {
+	return api.GenericUpdate(h, coord)
+}
 func (coord *TOCoordinate) Delete() (error, error, int) { return api.GenericDelete(coord) }
 
 func selectQuery() string {