+++ traffic_ops/traffic_ops_golang/deliveryservice/request/requests.go
@@ -44,6 +44,10 @@ type TODeliveryServiceRequest struct {
 	tc.DeliveryServiceRequestNullable
 }
 
+func (v *TODeliveryServiceRequest) GetLastUpdated() (*time.Time, bool, error) {
+	return api.GetLastUpdated(v.APIInfo().Tx, *v.ID, "deliveryservice_request")
+}
+
 func (v *TODeliveryServiceRequest) SetLastUpdated(t tc.TimeNoMod) { v.LastUpdated = &t }
 func (v *TODeliveryServiceRequest) InsertQuery() string           { return insertRequestQuery() }
 func (v *TODeliveryServiceRequest) UpdateQuery() string           { return updateRequestQuery() }
@@ -198,7 +202,7 @@ func (req TODeliveryServiceRequest) IsTenantAuthorized(user *auth.CurrentUser) (
 //ParsePQUniqueConstraintError is used to determine if a request with conflicting values exists
 //if so, it will return an errorType of DataConflict and the type should be appended to the
 //generic error message returned
-func (req *TODeliveryServiceRequest) Update() (error, error, int) {
+func (req *TODeliveryServiceRequest) Update(h http.Header) (error, error, int) {
 	if req.ID == nil {
 		return errors.New("missing id"), nil, http.StatusBadRequest
 	}
@@ -226,7 +230,7 @@ func (req *TODeliveryServiceRequest) Update() (error, error, int) {
 	userID := tc.IDNoMod(req.APIInfo().User.ID)
 	req.LastEditedByID = &userID
 
-	return api.GenericUpdate(req)
+	return api.GenericUpdate(h, req)
 }
 
 // Creator implements the tc.Creator interface
@@ -359,7 +363,7 @@ type deliveryServiceRequestAssignment struct {
 }
 
 // Update assignee only
-func (req *deliveryServiceRequestAssignment) Update() (error, error, int) {
+func (req *deliveryServiceRequestAssignment) Update(h http.Header) (error, error, int) {
 	// req represents the state the deliveryservice_request is to transition to
 	// we want to limit what changes here -- only assignee can change
 	if req.ID == nil {
@@ -382,6 +386,11 @@ func (req *deliveryServiceRequestAssignment) Update() (error, error, int) {
 	assigneeID := req.AssigneeID
 	req.DeliveryServiceRequestNullable = current.DeliveryServiceRequestNullable
 	req.AssigneeID = assigneeID
+
+	userErr, sysErr, statusCode := api.CheckIfUnModified(h, req.ReqInfo.Tx, *req.ID, "deliveryservice_request")
+	if userErr != nil || sysErr != nil {
+		return userErr, sysErr, statusCode
+	}
 
 	// LastEditedBy field should not change with status update
 	if _, err = req.APIInfo().Tx.Tx.Exec(`UPDATE deliveryservice_request SET assignee_id = $1 WHERE id = $2`, req.AssigneeID, *req.ID); err != nil {
@@ -422,7 +431,7 @@ type deliveryServiceRequestStatus struct {
 	TODeliveryServiceRequest
 }
 
-func (req *deliveryServiceRequestStatus) Update() (error, error, int) {
+func (req *deliveryServiceRequestStatus) Update(h http.Header) (error, error, int) {
 	// req represents the state the deliveryservice_request is to transition to
 	// we want to limit what changes here -- only status can change,  and only according to the established rules
 	// for status transition
@@ -446,6 +455,10 @@ func (req *deliveryServiceRequestStatus) Update() (error, error, int) {
 	req.Status = st
 
 	// LastEditedBy field should not change with status update
+	existingLastUpdated := current.LastUpdated
+	if !api.IsUnmodified(h, existingLastUpdated.Time) {
+		return errors.New("resource was modified"), nil, http.StatusPreconditionFailed
+	}
 
 	if _, err = req.APIInfo().Tx.Tx.Exec(`UPDATE deliveryservice_request SET status = $1 WHERE id = $2`, *req.Status, *req.ID); err != nil {
 		return api.ParseDBError(err)